Moved rfSchema, added docstrings
