+++ simpletypes.py
@@ -8,9 +8,16 @@ import traverseService as rst
 
 rsvLogger = rst.getLogger()
 
-def validateDeprecatedEnum(name, val, listEnum):
-    """
+def validateDeprecatedEnum(name: str, val, listEnum: list):
+    """validateDeprecatedEnum
+    
     Validates a DeprecatedEnum
+
+    :param name: Name of property (printing purposes) Property Name
+    :type name: str
+    :param val: Value for a given property
+    :param listEnum: List of expected enum values
+    :type listEnum: list
     """
     paramPass = True
     if isinstance(val, list):
@@ -32,7 +39,17 @@ def validateDeprecatedEnum(name, val, listEnum):
     return paramPass
 
 
-def validateEnum(name, val, listEnum):
+def validateEnum(name: str, val, listEnum: list):
+    """validateEnum
+
+    Validate an enum value
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
+    :param listEnum: List of expected enum values
+    :type listEnum: list
+    """
     paramPass = isinstance(val, str)
     if paramPass:
         paramPass = val in listEnum
@@ -43,9 +60,15 @@ def validateEnum(name, val, listEnum):
     return paramPass
 
 
-def validateString(name, val, pattern=None):
-    """
+def validateString(name: str, val, pattern=None):
+    """validateString
+
     Validates a string, given a value and a pattern
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
+    :param pattern: Regex pattern for value
     """
     paramPass = isinstance(val, str)
     if paramPass:
@@ -61,9 +84,14 @@ def validateString(name, val, pattern=None):
     return paramPass
 
 
-def validateDatetime(name, val):
-    """
+def validateDatetime(name: str, val):
+    """validateDatetime
+    
     Validates a Datetime, given a value (pattern predetermined)
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
     """
     paramPass = validateString(name, val, '.*(Z|(\+|-)[0-9][0-9]:[0-9][0-9])')
     if not paramPass:
@@ -71,9 +99,14 @@ def validateDatetime(name, val):
     return paramPass
 
 
-def validateGuid(name, val):
-    """
+def validateGuid(name: str, val):
+    """validateGuid
+
     Validates a Guid, given a value (pattern predetermined)
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
     """
     paramPass = validateString(name, val, "[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}")
     if not paramPass:
@@ -81,9 +114,16 @@ def validateGuid(name, val):
     return paramPass
 
 
-def validateInt(name, val, minVal=None, maxVal=None):
-    """
+def validateInt(name: str, val, minVal=None, maxVal=None):
+    """validateInt
+
     Validates a Int, then passes info to validateNumber
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
+    :param minVal: Minimum value (int or float)
+    :param maxVal: Maximum value (int or float)
     """
     if not isinstance(val, int):
         rsvLogger.error("{}: Expected integer, got type {}".format(name, str(type(val)).strip('<>')))
@@ -92,9 +132,16 @@ def validateInt(name, val, minVal=None, maxVal=None):
         return validateNumber(name, val, minVal, maxVal)
 
 
-def validateNumber(name, val, minVal=None, maxVal=None):
-    """
+def validateNumber(name: str, val, minVal=None, maxVal=None):
+    """validateNumber
+    
     Validates a Number and its min/max values
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
+    :param minVal: Minimum value (int or float)
+    :param maxVal: Maximum value (int or float)
     """
     paramPass = isinstance(val, (int, float))
     if paramPass:
@@ -111,9 +158,14 @@ def validateNumber(name, val, minVal=None, maxVal=None):
     return paramPass
 
 
-def validatePrimitive(name, val):
-    """
+def validatePrimitive(name: str, val):
+    """validatePrimitive
+
     Validates a Primitive
+
+    :param name: Name of property (printing purposes)
+    :type name: str
+    :param val: Value for a given property
     """
     if isinstance(val, (int, float, str, bool)):
         return True
